Intelligent mobile  robot
to construct a 2D map of the robot environment using only odometer and sonar sensors. In this experiment, the robot will be mapping the walls of
the room where it is able to detect the presence of obstacles using 8 sonar sensors. Every sensor will able to detect obstacle, read values and plot them. No prior knowledge of the environment is assumed, and the construction of map is depending on the robot autonomously. The use of sonar sensors with trigonometric and geometric approaches have made it possible to
identify most of the obstacles throughout its movement as normal localization and object detection are not completely error free in any robotic platform.