localization in 1D
To defining the robot position in a known environment with no information other than the robot is somewhere on the map. The Monte Carlo
Localization method for developing more efficient and more robust localization methods. The approach is to test using a mobile robot equipped with one
sonar sensor. The experiment result showed that the efficiency of localization is improved by actively controlling the robot motion direction with
actively controlling its sensors. The first section states the problem definition of this paper. The section two contains a detailed description of the
Monte-Carlo Simulation method i.e. particle filtering. Third section contains the software implementation part and last section demonstrates the results
of the algorithm.